requirements.txt
sounddevice
scipy
openai-whisper @ git+https://github.com/openai/whisper.git
openai
python-dotenv
torch==2.5.1
torchaudio==2.5.1
TTS
playsound==1.2.2
